Make functor in map_iterator_adapter a pointer
Used to be a reference. This allows to make the iterator default-constructible,
as required by the ForwardIterator concept if applicable. A possibly better
option would be to make it a std::optional<std::reference_wrapper<Functor> once
C++17 can be required. Even safer would be an optional shared pointer or similar
but `new` allocation is incompatible with lambdas.